complete Lc147
Again, be really familiar with the process of reversing a linkedlist first! and for here, whenever you meet a new element that is smaller than the current element, just put the new one in the front of the linkedlist, and do another loop to make the correct position for the new one. Because in a single directed linkedlist, it's much easier for us to move the node when we start from the head.